Fix mistake of assuming destination exists
Reworked the command running code to allow for errors, and leverage that to
allow for the destination dataset to not exist. Previously would crash when
the destination was not yet created (i.e. before the first replication).

Changed the test suite to expect a zpool to have been created and named via
an environment variable. Makes the test suite much less fragile.

rebar3 ct passes